pyproject.toml
[project]
name = "stable-diffusion-mini-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.33.4",
    "pillow==10.1.0",
    "pydantic==2.5.0",
    "pymongo>=4.13.2",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "streamlit>=1.46.1",
]
